test: create VideoParserClient.h
Add a common file for the VideoParserClient
Add a quiet mode.
