+++ test/test.cpp
@@ -17,164 +17,19 @@
 
 #include <glib.h>
 
-#include <atomic>
-#include <cerrno>
-#include <cstdio>
-#include <cstdlib>
-#include <cstring>
-#include <vector>
-#include <cinttypes>
 
-#include "dump.h"
 
-#include <vk_video/vulkan_video_codecs_common.h>
-#include <vulkan/vulkan_beta.h>
+
+#include "VideoParserClient.h"
 
 
 static VkVideoCodecOperationFlagBitsKHR codec = VK_VIDEO_CODEC_OPERATION_DECODE_H264_BIT_EXT;
 
-class PictureParameterSet : public VkParserVideoRefCountBase {
-public:
-    static PictureParameterSet* create()
-    {
-        return new PictureParameterSet();
-    }
 
-    int32_t AddRef() final
-    {
-        return ++m_refCount;
-    }
-
-    int32_t Release() final
-    {
-        uint32_t ret = --m_refCount;
-        // Destroy the device if refcount reaches zero
-        if (ret == 0)
-            delete this;
-        return ret;
-    }
-
-private:
-    std::atomic<int32_t> m_refCount;
-
-    PictureParameterSet()
-        : m_refCount(0)
-    {
-    }
-};
-
-class Picture : public VkPicIf {
-public:
-    Picture()
-        : m_refCount(0)
-    {
-    }
-
-    void AddRef() final
-    {
-        assert(m_refCount >= 0);
-        ++m_refCount;
-    }
-
-    void Release() final
-    {
-        assert(m_refCount > 0);
-        int32_t ref = --m_refCount;
-        if (ref == 0) {
-            decodeHeight = 0;
-            decodeWidth = 0;
-            decodeSuperResWidth = 0;
-        }
-    }
-
-    bool isAvailable() const { return m_refCount == 0; }
-
-private:
-    std::atomic<int32_t> m_refCount;
-};
-
-class VideoParserClient : public VkParserVideoDecodeClient {
-public:
-    VideoParserClient()
-        : m_dpb(32)
-    {
-    }
-
-    int32_t BeginSequence(const VkParserSequenceInfo* info) final
-    {
-        int32_t max = 16, conf = 1;
-
-        fprintf(stdout, "%s\n", __FUNCTION__);
-
-        dump_parser_sequence_info(info);
-
-        if (info->eCodec == VK_VIDEO_CODEC_OPERATION_ENCODE_H264_BIT_EXT)
-            max++;
-
-        if (info->nMinNumDecodeSurfaces > 0)
-            conf += info->nMinNumDecodeSurfaces - (info->isSVC ? 3 : 1);
-        if (conf > max)
-            conf = max;
-
-        return std::min(conf, 17);
-    }
-
-    bool AllocPictureBuffer(VkPicIf** pic) final
-    {
-        fprintf(stdout, "%s\n", __FUNCTION__);
-
-        for (auto& apic : m_dpb) {
-            if (apic.isAvailable()) {
-                apic.AddRef();
-                *pic = &apic;
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    bool DecodePicture(VkParserPictureData* pic) final
-    {
-        fprintf(stdout, "%s - %" PRIu32 "\n", __FUNCTION__, pic->nBitstreamDataLen);
-        dump_parser_picture_data(codec, pic);
-        return true;
-    }
-
-    bool UpdatePictureParameters(VkPictureParameters* params, VkSharedBaseObj<VkParserVideoRefCountBase>& shared, uint64_t count) final
-    {
-        // fprintf(stdout, "%s: %" PRIu64 "\n", __FUNCTION__, count);
-        fprintf(stdout, "%s\n", __FUNCTION__);
-        shared = PictureParameterSet::create();
-        dump_picture_parameters(params);
-        return true;
-    }
-
-    bool DisplayPicture(VkPicIf* pic, int64_t ts) final
-    {
-        fprintf(stdout, "%s\n", __FUNCTION__);
-        return true;
-    }
-
-    void UnhandledNALU(const uint8_t*, int32_t) final
-    {
-        fprintf(stdout, "%s\n", __FUNCTION__);
-    }
-
-    ~VideoParserClient()
-    {
-        for (auto& pic : m_dpb)
-            assert(pic.isAvailable());
-    }
-
-private:
-    std::vector<Picture> m_dpb;
-};
-
-static bool parse(FILE* stream)
+static bool parse(FILE* stream, bool quiet)
 {
     VulkanVideoDecodeParser* parser = nullptr;
-    VideoParserClient client = VideoParserClient();
+    VideoParserClient client = VideoParserClient(codec, quiet);
     VkParserInitDecodeParameters params = {
         .interfaceVersion = NV_VULKAN_VIDEO_PARSER_API_VERSION,
         .pClient = &client,
@@ -234,7 +89,7 @@ static bool parse(FILE* stream)
     return ret;
 }
 
-int process_file (gchar* filename) {
+int process_file (gchar* filename, bool quiet) {
     FILE* file;
     file = fopen(filename, "r");
     if (!file) {
@@ -242,7 +97,7 @@ int process_file (gchar* filename) {
         return EXIT_FAILURE;
     }
 
-    if (!parse(file)) {
+    if (!parse(file, quiet)) {
         fclose(file);
         return EXIT_FAILURE;
     }
@@ -257,9 +112,11 @@ int main(int argc, char** argv)
     GError *err = NULL;
     gchar **filenames = NULL;
     gchar *codec_str = NULL;
+    gboolean quiet = FALSE;
     gint ret = EXIT_SUCCESS;
 
     static GOptionEntry entries[] = {
+        { "quiet", 'q', 0, G_OPTION_ARG_NONE, &quiet, "Quiet parser", NULL },
         { "codec", 'c', 0, G_OPTION_ARG_STRING, &codec_str, "Codec to use ie h265", NULL },
         {G_OPTION_REMAINING, 0, 0, G_OPTION_ARG_FILENAME_ARRAY, &filenames, NULL},
         { NULL }
@@ -292,7 +149,7 @@ int main(int argc, char** argv)
       codec = VK_VIDEO_CODEC_OPERATION_DECODE_H265_BIT_EXT;
 
 
-    ret |= process_file (filenames[0]);
+    ret |= process_file (filenames[0], quiet);
 
     return ret;
 }
